requirements.txt
python-telegram-bot==20.7
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0 